requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-text-splitters>=0.3.0
openai>=1.0.0
langchain-openai>=0.0.5
pinecone>=5.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
docx2txt>=0.8
pandas>=2.0.0
tiktoken>=0.5.0

